CHE-5092 Code clean up
Move StartSynchronizer to DockerInternalRuntimes because it is unused in DockerRuntimeContext
Fix destroying of a newly created machine when start is canceled
Move publishing of STARTING workspace event to a place where runtime is already ready
